requirements.txt
Django==1.3.1
PIL==1.1.7
South==0.7.5
argparse==1.2.1
distribute==0.6.24
django-classy-tags==0.3.4.1
django-cms==2.2
django-mptt==0.5.2
django-photologue==2.3
django-picklefield==0.2.1
django-sekizai==0.6.1
django-tagging==0.3.1
html5lib==0.95
wsgiref==0.1.2
